+++ src/main/resources/application-local.yml
@@ -1,7 +1,7 @@
 spring:
   mail:
-    username: // 개발자 gmail
-    password: // 개발자 본인의 앱 비밀번호
+    username: talent8661@gmail.com
+    password: unxs tjgs ftcc jhlp
   
   data:
     redis:
